modernized the CGI part of the source code
