+++ cmake-static-release.sh
@@ -1,14 +1,9 @@
 #!/usr/bin/env bash
 set -e
 
-cores=$(grep -c processor /proc/cpuinfo)
-
-mkdir -p cmake-build-release
-pushd cmake-build-release
-cmake -DCMAKE_BUILD_TYPE=Release -DCMAKE_BUILD_STATIC=1 ../src
-make -j$cores
+BUILD=build-release
+cmake -DCMAKE_BUILD_TYPE=Release -DCMAKE_BUILD_STATIC=ON -B "$BUILD" src
+cmake --build "$BUILD"
 echo
 echo "Success:"
-file iben
-popd
-
+file "$BUILD/iben"
