+++ src/iben.hpp
@@ -20,45 +20,45 @@
 #ifndef __IBEN_H_
 #define __IBEN_H_
 
-#include <map>
-#include <string>
-
 #include <bdd.h>
+#include <map>
 #include <string.h>
+#include <string>
 
 #define PSV "gsview"
 
-#define EXPR 1
+#define EXPR     1
 #define VARASSOC 2
-#define NEW 3
-#define VAR 4
+#define NEW      3
+#define VAR      4
 
 struct st_entry
 {
     int type;
     std::string id;
-    struct {
-	struct
-	{
-	    bdd f;
-	} expr;
-	struct
-	{
-	    bddPair *p;
-	} varassoc;
-	int var;
+    struct
+    {
+        struct
+        {
+            bdd f;
+        } expr;
+        struct
+        {
+            bddPair* p;
+        } varassoc;
+        int var;
     } info;
-  ~st_entry();
-  void free();
+    ~st_entry();
+    void free();
 };
 
 extern bool terminated;
 extern bool interactive;
 extern bool timer_running;
 extern int line_number;
-extern std::map<std::string,st_entry> symbol_table;
-extern std::map<int,std::string> name_table;
-extern std::map<std::string,int> keywords;
+extern std::map<std::string, st_entry> symbol_table;
+extern std::map<int, std::string> name_table;
+extern std::map<std::string, int> keywords;
 extern double user_start;
 extern double system_start;
 extern double user_last;
@@ -66,12 +66,12 @@ extern double system_last;
 
 extern void printdot(FILE*, bdd);
 extern void print_sat(bdd);
-extern const char *undefined(const char *);
+extern const char* undefined(const char*);
 
-extern void get_cpu_time(double *, double *);
+extern void get_cpu_time(double*, double*);
 extern int yyparse();
-extern void parse(const char *);
-extern void yyerror(const char *);
+extern void parse(const char*);
+extern void yyerror(const char*);
 extern void print_version();
 
 #endif
